Clear warning

+++ src/MPR121.cpp
@@ -233,7 +233,7 @@ uint8_t MPR121_Class::readRegister8(uint8_t reg)
         _wire->beginTransmission(_i2caddr);
         _wire->write(reg);
         _wire->endTransmission(false);
-        _wire->requestFrom(_i2caddr, 1);
+        _wire->requestFrom(_i2caddr, (uint8_t)1);
         if (_wire->available() < 1)
             return 0;
         return (_wire->read());
@@ -258,7 +258,7 @@ uint16_t MPR121_Class::readRegister16(uint8_t reg)
         _wire->beginTransmission(_i2caddr);
         _wire->write(reg);
         _wire->endTransmission(false);
-        _wire->requestFrom(_i2caddr, 2);
+        _wire->requestFrom(_i2caddr, (uint8_t)2);
         if (_wire->available() < 2)
             return 0;
         v = _wire->read();
